Enable open link in browser if not direct download
For some people if can produce json (can do it in github raw) but cant host ROM in personal server
and host the ROM in filesharing like AFH or Basketbuild or Dev-host or anything else, you can use it now.
The default json from DU is just filename, filesize(optional) and downloads, to use this add attribute direct.

Now the json is like this:
{"dev_info":
[
 {
  "filename":"file_name.zip",
  "filesize":303955562,
  "downloads":"http:\/\/www.example.com\/file_name.zip",
  "direct":true
 }
]}

Set "direct" to false if want to open in browser (host in filesharing).

Change-Id: I28a048f0ae35d336fb7057b71441269d16bace6f